requirements.txt
flask
requests
numpy
pyjwt